requirements.txt
Django>=4.2
Pillow
opencv-python
python-lzw
python-arithmetic-coding
